automation-test_new/requirements.txt
xlrd
django-suit
certifi==2018.11.29
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
idna==2.8
pytz==2018.9
requests==2.21.0
django-rest-swagger==2.2.0
python-docx==0.8.10
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
python-crontab==2.3.6
PyMySQL==0.9.3
XlsxWriter==1.1.2
python-docx==0.8.10
django-multiselectfield==0.1.8
xpinyin==0.5.6
dubbo_telnet
jsonschema